Se corrige error de poder ingresar desde author a posts de admin(permisos)

+++ app/Http/Controllers/Author/PostController.php
@@ -108,6 +108,12 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if($post->user_id != Auth::id())
+        {
+            Toastr::error('You are not authorized to access this post','Error');
+            return redirect()->back();
+        }
+        
         return view('author.post.show', compact('post'));
     }
 
@@ -119,6 +125,12 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if($post->user_id != Auth::id())
+        {
+            Toastr::error('You are not authorized to access this post','Error');
+            return redirect()->back();
+        }
+
         $categories = Category::all();
         $tags = Tag::all();
         return view('author.post.edit',compact('post','categories','tags'));
@@ -133,6 +145,12 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if($post->user_id != Auth::id())
+        {
+            Toastr::error('You are not authorized to access this post','Error');
+            return redirect()->back();
+        }
+
         $this->validate($request,[
             'title' => 'required',
             'image' => 'image',
@@ -198,6 +216,12 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if($post->user_id != Auth::id())
+        {
+            Toastr::error('You are not authorized to access this post','Error');
+            return redirect()->back();
+        }
+        
         if(Storage::disk('public')->exists('post/'.$post->image))
         {
             Storage::disk('public')->delete('post/'.$post->image);
